refactor: move window to own file

+++ src/input.rs
@@ -3,7 +3,7 @@ use vek::Vec2;
 /// Current input.
 #[derive(Debug, Default)]
 pub struct Input {
-    pub mouse_pos: Vec2<f64>,
+    pub mouse_pos: Vec2<i32>,
     pub up_pressed: bool,
     pub down_pressed: bool,
     pub left_pressed: bool,
